【Fix】Fatモデルへの対応｜計算ロジックを自作モジュールへ移行
